Remove CameraAnimationsManagerProtocol.mapView
