+++ Sources/MapboxMaps/Gestures/CameraAnimationsManagerProtocol.swift
@@ -2,8 +2,6 @@ import Foundation
 
 internal protocol CameraAnimationsManagerProtocol: AnyObject {
 
-    var mapView: MapView? { get }
-
     var options: CameraBoundsOptions { get }
 
     func ease(to camera: CameraOptions,
